+++ sdk/resourcemanager/dataprotection/armdataprotection/zz_generated_operationresult_client.go
@@ -93,7 +93,7 @@ func (client *OperationResultClient) getCreateRequest(ctx context.Context, opera
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-02-01-preview")
+	reqQP.Set("api-version", "2022-03-31-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header.Set("Accept", "application/json")
 	return req, nil