+++ sdk/resourcemanager/dataprotection/armdataprotection/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Release History
 
+## 0.5.0 (2022-04-26)
+### Features Added
+
+- New const `AutoHealStatusOff`
+- New const `AutoHealStatusOn`
+- New function `*KubernetesPVRestoreCriteria.UnmarshalJSON([]byte) error`
+- New function `*KubernetesPVRestoreCriteria.GetItemLevelRestoreCriteria() *ItemLevelRestoreCriteria`
+- New function `KubernetesStorageClassRestoreCriteria.MarshalJSON() ([]byte, error)`
+- New function `*KubernetesStorageClassRestoreCriteria.UnmarshalJSON([]byte) error`
+- New function `*KubernetesStorageClassRestoreCriteria.GetItemLevelRestoreCriteria() *ItemLevelRestoreCriteria`
+- New function `PossibleAutoHealStatusValues() []AutoHealStatus`
+- New function `KubernetesPVRestoreCriteria.MarshalJSON() ([]byte, error)`
+- New struct `AutoHealSettings`
+- New struct `BackupInstancesExtensionRoutingClientListOptions`
+- New struct `BackupInstancesExtensionRoutingClientListResponse`
+- New struct `KubernetesPVRestoreCriteria`
+- New struct `KubernetesStorageClassRestoreCriteria`
+
+
 ## 0.4.0 (2022-04-15)
 ### Breaking Changes